Pausing clearScreen during editing
