+++ packages/core/README.md
@@ -28,18 +28,17 @@ __0.8.5__
 ## Test coverage
 
 ```
-  92 passing (2s)
+  62 passing (2s)
 
 -----------------|---------|----------|---------|---------|-------------------
 File             | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
 -----------------|---------|----------|---------|---------|-------------------
-All files        |   99.82 |    95.29 |     100 |   99.82 |                   
- src             |    99.8 |       96 |     100 |    99.8 |                   
-  Crypto.js      |   99.42 |    90.32 |     100 |    99.4 | 266               
+All files        |     100 |    95.98 |     100 |     100 |                   
+ src             |     100 |    96.97 |     100 |     100 |                   
+  Crypto.js      |     100 |       80 |     100 |     100 | 17-21             
   Entry.js       |     100 |    95.65 |     100 |     100 | 40                
   Secrez.js      |     100 |    98.54 |     100 |     100 | 111,438           
   _Secrez.js     |     100 |    96.43 |     100 |     100 | 124               
-  index.js       |     100 |      100 |     100 |     100 |                   
  src/config      |     100 |    88.46 |     100 |     100 |                   
   ConfigUtils.js |     100 |    88.46 |     100 |     100 | 82-87,111         
   index.js       |     100 |      100 |     100 |     100 |
