+++ packages/core/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/core",
-  "version": "0.8.8",
+  "version": "0.8.9",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -10,8 +13,8 @@
     "posttest": "nyc check-coverage --statements 99 --branches 95 --functions 99 --lines 99"
   },
   "dependencies": {
-    "@secrez/crypto": "workspace:~0.1.0",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/crypto": "workspace:~0.1.3",
+    "@secrez/utils": "workspace:~0.1.6",
     "base-x": "^3.0.8",
     "bip39": "^3.0.2",
     "fs-extra": "^8.1.0",
@@ -24,7 +27,7 @@
     "tweetnacl-util": "^0.15.1"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
