+++ packages/courier/README.md
@@ -50,7 +50,18 @@ The connected Secrez account will then be able to receive message in real time o
 ## Test coverage
 
 ```
+  12 passing (7s)
 
+------------|---------|----------|---------|---------|----------------------------------------------
+File        | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s                            
+------------|---------|----------|---------|---------|----------------------------------------------
+All files   |   79.44 |    63.87 |   79.55 |   79.35 |                                              
+ App.js     |   79.35 |    60.98 |   69.23 |   79.35 | ...0,145,170,175-177,181-182,191-193,205,215 
+ Config.js  |     100 |      100 |     100 |     100 |                                              
+ Courier.js |   83.33 |        0 |      50 |   83.33 | 12                                           
+ Db.js      |    91.3 |    74.19 |     100 |    91.3 | 82-85                                        
+ Server.js  |   69.66 |     56.1 |   71.43 |   69.32 | 16,56-63,93,96,113,137-153,158-162,167-178   
+------------|---------|----------|---------|---------|----------------------------------------------
 ```
 ## Copyright
