+++ packages/courier/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/courier",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'bin/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -13,12 +16,12 @@
     "secrez-courier": "bin/secrez-courier.js"
   },
   "dependencies": {
-    "@secrez/core": "workspace:~0.8.8",
-    "@secrez/fs": "workspace:~0.8.4",
-    "@secrez/hub": "workspace:~0.1.5",
-    "@secrez/tls": "workspace:~0.1.3",
-    "@secrez/tunnel": "workspace:~0.1.3",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/core": "workspace:~0.8.9",
+    "@secrez/fs": "workspace:~0.8.5",
+    "@secrez/hub": "workspace:~0.1.6",
+    "@secrez/tls": "workspace:~0.1.4",
+    "@secrez/tunnel": "workspace:~0.1.4",
+    "@secrez/utils": "workspace:~0.1.6",
     "body-parser": "^1.19.0",
     "chalk": "^3.0.0",
     "command-line-args": "^5.1.1",
@@ -33,7 +36,7 @@
     "superagent": "^5.3.1"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
