+++ packages/crypto/README.md
@@ -2,7 +2,9 @@
 
 Secrez is the secrets manager for the cryptocurrencies era.
 
-@secrez/crypto is the basic crypto library.
+@secrez/crypto is the basic crypto library. It includes only functions that can be called also in a browser. 
+
+In @secrez/core it is integrated with methods who works only server side, in order to be used in Secrez.
 
 ## TODO
 
@@ -10,21 +12,24 @@ API documentation
 
 ## History
 
+__0.1.2__
+* moving all methods that don't work in a browser back to @secrez/core
+
 __0.1.0__
-* moved the package from @secrez/core to a separate libr 
+* moved from @secrez/core to this separate library 
 
 
 ## Test coverage
 
 ```
-  32 passing (654ms)
+  31 passing (713ms)
 
 ----------|---------|----------|---------|---------|-------------------
 File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
 ----------|---------|----------|---------|---------|-------------------
-All files |     100 |    92.59 |     100 |     100 |                   
- index.js |     100 |    92.59 |     100 |     100 | 28-35,72,132      
-----------|---------|----------|---------|---------|------------------- 
+All files |     100 |    92.31 |     100 |     100 |                   
+ index.js |     100 |    92.31 |     100 |     100 | 27-34,71,112      
+----------|---------|----------|---------|---------|-------------------
 ```
