+++ packages/crypto/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/crypto",
-  "version": "0.1.0",
+  "version": "0.1.3",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -12,14 +15,13 @@
   "dependencies": {
     "base-x": "^3.0.8",
     "bip39": "^3.0.2",
-    "microtime": "^3.0.0",
     "sha3": "^2.1.2",
     "shamir": "^0.7.1",
     "tweetnacl": "^1.0.3",
     "tweetnacl-util": "^0.15.1"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/utils": "workspace:~0.1.6",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
