+++ packages/fs/README.md
@@ -21,7 +21,21 @@ API documentation
 ## Test coverage
 
 ```
+  96 passing (3s)
 
+---------------|---------|----------|---------|---------|-------------------------------------------
+File           | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s                         
+---------------|---------|----------|---------|---------|-------------------------------------------
+All files      |    91.4 |    80.57 |   92.54 |    91.3 |                                           
+ DataCache.js  |   89.41 |    74.07 |   90.91 |   89.29 | 18-24,123-124,150                         
+ ExternalFs.js |   94.03 |    76.92 |     100 |   93.94 | 10,28,34,50                               
+ FsUtils.js    |     100 |      100 |     100 |     100 |                                           
+ InternalFs.js |   86.97 |    72.34 |    91.3 |   86.86 | ...48,265,274,312,322,348,404-409,427-431 
+ Messages.js   |     100 |      100 |     100 |     100 |                                           
+ Node.js       |   94.44 |    88.89 |   95.83 |   94.32 | ...41,519,537-543,603,609,624-628,730,738 
+ Tree.js       |   90.16 |    74.42 |   86.11 |   90.11 | ...72,418-441,472,482,533,625,628,655-664 
+ index.js      |     100 |      100 |     100 |     100 |                                           
+---------------|---------|----------|---------|---------|-------------------------------------------
 ```
