+++ packages/fs/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/fs",
-  "version": "0.8.4",
+  "version": "0.8.5",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -10,14 +13,14 @@
     "posttest": "nyc check-coverage --statements 85 --branches 80 --functions 85 --lines 85"
   },
   "dependencies": {
-    "@secrez/core": "workspace:~0.8.8",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/core": "workspace:~0.8.9",
+    "@secrez/utils": "workspace:~0.1.6",
     "command-line-args": "^5.1.1",
     "fs-extra": "^8.1.0",
     "lodash": "^4.17.15"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
