+++ packages/hub/README.md
@@ -66,7 +66,22 @@ When I have a moment, I will write a post on how to do it using Let's Encrypt.
 
 ## Test coverage
 ```
+  5 passing (179ms)
 
+-------------------|---------|----------|---------|---------|---------------------------------------
+File               | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s                     
+-------------------|---------|----------|---------|---------|---------------------------------------
+All files          |   73.48 |    60.61 |   70.69 |   73.42 |                                       
+ src               |   78.16 |    68.24 |   90.48 |   78.16 |                                       
+  Validator.js     |      80 |       50 |     100 |      80 | 28-30,34                              
+  createServer.js  |   76.55 |    69.57 |      90 |   76.55 | ...04,216-228,235-236,241-242,248-249 
+  utils.js         |   82.93 |     69.7 |   85.71 |   82.93 | 12,19-23,42,53                        
+ src/lib           |   68.42 |    46.81 |   59.46 |   68.25 |                                       
+  Client.js        |   60.71 |       40 |   42.86 |      60 | 21,37,44,53-83,91-94,101-102,107-109  
+  ClientManager.js |   70.27 |     37.5 |   71.43 |   70.27 | 35,57,70-84                           
+  Shortener.js     |     100 |    83.33 |     100 |     100 | 20                                    
+  TunnelAgent.js   |   65.82 |    43.48 |   61.54 |   65.82 | ...19,132-136,148-149,160-163,171-172 
+-------------------|---------|----------|---------|---------|---------------------------------------
 ```
 
 ## Copyright
