+++ packages/hub/package.json
@@ -1,8 +1,11 @@
 {
   "name": "@secrez/hub",
   "description": "A hub end-to-end encrypted communication",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "repository": {
     "type": "git",
     "url": "git://github.com/secrez/secrez.git"
@@ -22,9 +25,9 @@
   },
   "preferGlobal": true,
   "dependencies": {
-    "@secrez/core": "workspace:~0.8.8",
-    "@secrez/tls": "workspace:~0.1.3",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/core": "workspace:~0.8.9",
+    "@secrez/tls": "workspace:~0.1.4",
+    "@secrez/utils": "workspace:~0.1.6",
     "book": "1.3.3",
     "chalk": "^3.0.0",
     "command-line-args": "^5.1.1",
@@ -34,7 +37,7 @@
     "pump": "3.0.0"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "4.2.0",
     "cross-env": "7.0.2",
     "eslint": "^6.8.0",
