+++ packages/secrez/package.json
@@ -1,6 +1,6 @@
 {
   "name": "secrez",
-  "version": "0.8.8",
+  "version": "0.8.9",
   "license": "MIT",
   "scripts": {
     "dev": "cross-env NODE_ENV=dev bin/secrez.js -c `pwd`/tmp/secrez-dev -i 1e3 -l `pwd`/tmp",
@@ -15,10 +15,10 @@
     "exclude": []
   },
   "dependencies": {
-    "@secrez/core": "workspace:~0.8.8",
-    "@secrez/fs": "workspace:~0.8.4",
-    "@secrez/hub": "workspace:~0.1.5",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/core": "workspace:~0.8.9",
+    "@secrez/fs": "workspace:~0.8.5",
+    "@secrez/hub": "workspace:~0.1.6",
+    "@secrez/utils": "workspace:~0.1.6",
     "case": "^1.6.3",
     "chalk": "^3.0.0",
     "clear": "^0.1.0",
@@ -37,8 +37,8 @@
     "tiny-cli-editor": "^0.1.1"
   },
   "devDependencies": {
-    "@secrez/courier": "workspace:~0.1.4",
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/courier": "workspace:~0.1.5",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "chalk": "^3.0.0",
     "cross-env": "^7.0.2",
