+++ packages/secrez/src/commands/Edit.js
@@ -155,6 +155,7 @@ class Edit extends require('../Command') {
     if (options.help) {
       return this.showHelp()
     }
+    this.prompt.clearScreen.pause(true)
     let currentEditor
     try {
       this.validate(options, {
@@ -172,6 +173,7 @@ class Edit extends require('../Command') {
     }
     // eslint-disable-next-line require-atomic-updates
     process.env.EDITOR = currentEditor
+    this.prompt.clearScreen.pause(true)
     await this.prompt.run()
   }
 }
