+++ packages/test-helpers/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/test-helpers",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -10,10 +13,10 @@
     "posttest": "nyc check-coverage --statements 100 --branches 90 --functions 100 --lines 100"
   },
   "dependencies": {
-    "@secrez/core": "workspace:~0.8.8",
-    "@secrez/fs": "workspace:~0.8.4",
-    "@secrez/hub": "workspace:~0.1.5",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/core": "workspace:~0.8.9",
+    "@secrez/fs": "workspace:~0.8.5",
+    "@secrez/hub": "workspace:~0.1.6",
+    "@secrez/utils": "workspace:~0.1.6",
     "chai": "^4.2.0",
     "chalk": "^3.0.0",
     "fs-extra": "^8.1.0",
