+++ packages/tls/README.md
@@ -8,7 +8,14 @@ A documentation will come soon.
 ## Test coverage
 
 ```
+  4 passing (582ms)
 
+----------|---------|----------|---------|---------|-------------------
+File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
+----------|---------|----------|---------|---------|-------------------
+All files |     100 |    92.86 |     100 |     100 |                   
+ TLS.js   |     100 |    92.86 |     100 |     100 | 8                 
+----------|---------|----------|---------|---------|-------------------
 ```
 
 ## Copyright
