+++ packages/tls/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/tls",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'bin/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -10,7 +13,7 @@
     "posttest": "nyc check-coverage --statements 99 --branches 90 --functions 99 --lines 99"
   },
   "dependencies": {
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/utils": "workspace:~0.1.6",
     "chalk": "^4.1.0",
     "command-line-args": "^5.1.1",
     "cross-env": "^7.0.2",
@@ -19,7 +22,7 @@
     "nyc": "^15.1.0"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
