+++ packages/tunnel/README.md
@@ -9,7 +9,19 @@ It is used internally by the @secrez/courier and it is not supposed to be used d
 
 ## Test coverage
 ```
+  5 passing (880ms)
 
+---------------------------|---------|----------|---------|---------|-------------------------------
+File                       | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s             
+---------------------------|---------|----------|---------|---------|-------------------------------
+All files                  |   80.49 |    48.48 |   83.33 |   80.75 |                               
+ tunnel                    |      75 |       40 |      75 |      80 |                               
+  index.js                 |      75 |       40 |      75 |      80 | 8-9                           
+ tunnel/src                |   80.92 |       50 |   84.38 |   80.79 |                               
+  HeaderHostTransformer.js |     100 |       40 |     100 |     100 | 4-11                          
+  Tunnel.js                |   81.58 |    38.46 |   88.24 |   81.33 | ...08-109,124,154-155,162,170 
+  TunnelCluster.js         |   77.61 |       64 |      75 |   77.61 | 50-65,71-73,80,109-119        
+---------------------------|---------|----------|---------|---------|-------------------------------
 ```
 
 ## Copyright
