+++ packages/tunnel/package.json
@@ -1,8 +1,11 @@
 {
   "name": "@secrez/tunnel",
   "description": "Expose localhost for secure messaging",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "repository": {
     "type": "git",
     "url": "git://github.com/secrez/secrez.git"
@@ -27,8 +30,8 @@
     "posttest": "nyc check-coverage --statements 80 --branches 45 --functions 80 --lines 80"
   },
   "dependencies": {
-    "@secrez/core": "workspace:~0.8.8",
-    "@secrez/utils": "workspace:~0.1.5",
+    "@secrez/core": "workspace:~0.8.9",
+    "@secrez/utils": "workspace:~0.1.6",
     "chalk": "^3.0.0",
     "command-line-args": "^5.1.1",
     "fs-extra": "^9.0.1",
@@ -36,9 +39,9 @@
     "superagent": "^5.3.1"
   },
   "devDependencies": {
-    "@secrez/hub": "workspace:~0.1.5",
-    "@secrez/test-helpers": "workspace:~0.1.3",
-    "@secrez/tls": "workspace:~0.1.3",
+    "@secrez/hub": "workspace:~0.1.6",
+    "@secrez/test-helpers": "workspace:~0.1.4",
+    "@secrez/tls": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
