+++ packages/utils/README.md
@@ -8,7 +8,15 @@ This is a utils library used by other packages in the @secrez suite.
 ## Test coverage
 
 ```
+  37 passing (170ms)
 
+-------------|---------|----------|---------|---------|-------------------
+File         | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
+-------------|---------|----------|---------|---------|-------------------
+All files    |   94.38 |    84.52 |   86.84 |   94.22 |                   
+ UglyDate.js |     100 |    96.67 |     100 |     100 | 61                
+ index.js    |   92.19 |    77.78 |   85.71 |   91.87 | 233-255           
+-------------|---------|----------|---------|---------|-------------------
 ```
 
 ## Copyright
