+++ packages/utils/package.json
@@ -1,7 +1,10 @@
 {
   "name": "@secrez/utils",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "lint": "eslint -c .eslintrc 'src/**/*.js' 'test/**/*.js'",
     "all-tests": "find test/** -name '*.test.js' | xargs ./node_modules/.bin/mocha -R spec",
@@ -20,7 +23,7 @@
     "yaml": "^1.10.0"
   },
   "devDependencies": {
-    "@secrez/test-helpers": "workspace:~0.1.3",
+    "@secrez/test-helpers": "workspace:~0.1.4",
     "chai": "^4.2.0",
     "chalk": "^3.0.0",
     "cross-env": "^7.0.2",
